==> code/1_abre_teach.R <==

# Abre a base do TEACH e faz os reshapes
# Vitor Pereira, Set 2022

rm(list=ls()) # apaga tudo no ambiente

pacman:: p_load(dplyr, readxl, tidyr, 
                janitor)

# Paths
root     <- "C:/Users/vitor/Dropbox (Personal)/Sao Tome e Principe/2022/TEACH_STP2/"
input    <- paste0(root, "input/")
output   <- paste0(root, "output/")
tmp      <- paste0(root, "tmp/")
code     <- paste0(root, "code/")
setwd(root)

# Hora de STP
Sys.setenv(TZ="UCT")

######################################

base_teach <- read_excel(paste0(input,
      "Teach_STP_2022_v2.xlsx")) 

base_teach2 <- base_teach %>% 
rename (
  escola											= por_favor_selecione_a_escola_na_lista																											,							
  data											= data_em_que_a_aula_ocorreu																													,	
  classe											= classe																																		,	
  professor										= por_favor_selecione_nome_do_a_professor_a_na_lista																							,	
  observador										= por_favor_selecione_o_a_observador_a																											,
  observador_outro								= por_favor_especifique_nome_do_a_observador_a																									,
  professor_presente								= o_professor_titular_esta_presente																												,
  professor_sexo									= sexo_do_professor_titular																														,
  professor_suplente								= se_o_professor_titular_faltou_tem_algum_professor_suplente																					,	
  professor_suplente_sexo							= sexo_do_professor_suplente																													,	
  professor_idade									= qual_e_a_sua_idade																															,	
  professor_nivel									= qual_e_o_nivel_de_formacao_academica_concluida																								,	
  professor_experiencia							= ha_quantos_anos_voce_trabalha_como_professor_nessa_escola																						,
  professor_agregado_tamanho						= qual_e_o_seu_agregado_familiar																												,	
  opiniao_prof1 									= a_inteligencia_e_algo_inato_que_nao_pode_ser_profundamente_modificado																			,
  opiniao_prof2 									= pode_se_aprender_coisas_novas_mas_nao_se_pode_mudar_o_nivel_de_inteligencia_que_uma_pessoa_possui												,
  opiniao_prof3 									= ter_um_bom_desempenho_na_tarefa_e_uma_boa_maneira_de_o_estudante_mostrar_aos_outros_que_e_inteligente                           				,
  opiniao_prof4 									= um_estudante_se_sai_mal_numa_tarefa_eu_questiono_sua_inteligencia_ou_aptidao                                                    				,
  opiniao_prof5 									= quando_o_aluno_faz_muito_esforco_para_aprender_algo_ele_revela_que_nao_e_muito_inteligente                                      				,
  opiniao_prof6 									= preparar_se_bem_antes_de_enfrentar_uma_tarefa_e_uma_maneira_do_estudante_desenvolver_sua_propria_inteligencia                   				,
  opiniao_prof7 									= os_rapazes_terao_mais_sucesso_que_as_raparigas_na_vida                                                                          				,
  opiniao_prof8 									= diante_das_dificuldades_desta_escola_um_pequeno_aprendizado_dos_alunos_ja_e_um_bom_resultado                                    				,
  opiniao_prof9 									= com_as_familias_que_os_alunos_dessa_escola_tem_a_possibilidade_de_aprendizagem_fica_muito_comprometida                          				,
  opiniao_prof10									= a_quantidade_dos_alunos_por_turma_e_a_principal_razao_que_a_maioria_dos_alunos_nao_aprendem                                     				,
  opiniao_prof11									= os_alunos_dessa_escola_nao_tem_vontade_de_aprender                                                                              				,
  expect_prof_1									= saberao_ler_um_pequeno_texto_simples_ate_o_final_do_ano_lectivo																				,	
  expect_prof_2									= saberao_escrever_frases_simples_ate_o_final_do_ano_lectivo																					,	
  expect_prof_3									= saberao_somar_usando_numeros_ate_10_ate_o_final_do_ano_lectivo                                                                  				,
  expect_prof_4									= apenas_para_os_alunos_da_6a_classe_irao_ser_aprovados_no_exame_nacional_da_6a_classe                                            				,
  expect_prof_5									= apenas_para_os_alunos_da_6a_classe_terminarao_a_9a_classe                                                                       				,
  duracao_seg1									= segmento_1_duracao																															,	
  tempo_1_seg1									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_52															,
  tempo_1_envolvidos_seg1							= x0_2_os_alunos_estao_envolvidos_na_tarefa_53																									,
  tempo_2_seg1									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_54															,
  tempo_2_envolvidos_seg1							= x0_2_os_alunos_estao_envolvidos_na_tarefa_55																									,
  tempo_3_seg1									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_56															,
  tempo_3_envolvidos_seg1							= x0_2_os_alunos_estao_envolvidos_na_tarefa_57                                                                                    				,
  a1_apoio_aprend_1_seg1							= x1_1_a_professora_trata_todos_os_alunos_com_respeito_61																						,	
  a1_apoio_aprend_2_seg1							= x1_2_a_professora_usa_uma_linguagem_positiva_com_os_alunos_62																					,
  a1_apoio_aprend_3_seg1							= x1_3_a_professora_corresponde_as_necessidades_dos_alunos_63                                                                     				,
  a1_apoio_aprend_4_seg1							= x1_4_a_professora_nao_apresenta_preconceitos_de_genero_e_rejeita_estereotipos_de_genero_na_sala_de_aula_64                      				,
  a1_expect_posit_1_seg1							= x2_1_a_professora_define_claramente_as_expetativas_de_comportamento_para_as_atividades_na_sala_de_aula_75										,
  a1_expect_posit_2_seg1							= x2_2_a_professora_reconhece_o_comportamento_positivo_dos_alunos_76																			,	
  a1_expect_posit_3_seg1							= x2_3_a_professora_redireciona_o_mau_comportamento_e_concentra_se_no_comportamento_esperado_em_vez_de_no_comportamento_indesejado_77     		,
  a2_licao_clara_1_seg1							= x3_1_a_professora_articula_explicitamente_os_objetivos_da_aula_e_relaciona_as_atividades_da_turma_com_os_objetivos_88							,
  a2_licao_clara_2_seg1							= x3_2_a_explicacao_da_professora_sobre_o_conteudo_e_clara_89																					,	
  a2_licao_clara_3_seg1							= x3_3_durante_a_aula_a_professora_estabelece_conexoes_e_relaciona_as_com_outros_conteudos_conhecidos_ou_com_a_vida_quotidiana_dos_alunos_90	,	
  a2_licao_clara_4_seg1							= x3_4_a_professora_exemplifica_demonstrando_ou_pensando_em_voz_alta_91																			,
  a2_compreens_1_seg1								= x4_1_a_professora_faz_perguntas_da_pistas_ou_usa_outras_estrategias_para_determinar_qual_e_o_nivel_de_compreensao_dos_alunos_102				,
  a2_compreens_2_seg1								= x4_2_a_professora_monitoriza_a_maioria_dos_alunos_durante_o_trabalho_individual_ou_de_grupo_103												,	
  a2_compreens_3_seg1								= x4_3_a_professora_ajusta_o_ensino_ao_nivel_dos_alunos_104                                                                                   	,
  a2_coment_const_1_seg1							= x5_1_a_professora_faz_comentarios_especificos_ou_da_pistas_que_ajudam_a_esclarecer_os_equivocos_dos_alunos_115								,	
  a2_coment_const_2_seg1							= x5_2_a_professora_comentarios_especificos_ou_da_pistas_que_ajudam_os_alunos_a_identificarem_os_seus_sucessos_116								,
  a2_racioc_crit_1_seg1							= x6_1_a_professora_faz_perguntas_de_resposta_aberta_127																						,	
  a2_racioc_crit_2_seg1							= x6_2_a_professora_propoe_atividades_de_raciocinio_128																							,
  a2_racioc_crit_3_seg1							= x6_3_os_alunos_fazem_perguntas_de_resposta_aberta_ou_desenvolvem_atividades_de_raciocinio_129                                   				,
  a3_autonomia_1_seg1								= x7_1_a_professora_proporciona_escolhas_aos_alunos_140																							,
  a3_autonomia_2_seg1								= x7_2_a_professora_oferece_oportunidades_aos_alunos_para_assumirem_papeis_na_sala_de_aula_141													,
  a3_autonomia_3_seg1								= x7_3_os_alunos_voluntariam_se_para_participar_na_aula_142                                           											,
  a3_persev_1_seg1								= x8_1_a_professora_reconhece_o_esforco_dos_alunos_153																							,
  a3_persev_2_seg1								= x8_2_a_professora_tem_uma_atitude_positiva_em_relacao_aos_desafios_dos_alunos_154																,
  a3_persev_3_seg1								= x8_3_a_professora_incentiva_a_definicao_de_objetivos_155                                    													,
  a3_cap_soc_colab_1_seg1							= x9_1_a_professora_promove_a_colaboracao_entre_alunos_atraves_da_interacao_entre_colegas_166													,	
  a3_cap_soc_colab_2_seg1							= x9_2_a_professora_promove_as_capacidades_interpessoais_dos_alunos_167                           												,
  a3_cap_soc_colab_3_seg1							= x9_3_os_alunos_colaboram_uns_com_os_outros_atraves_da_interacao_entre_eles_168                  												,
  duracao_seg2									= segmento_2_duracao	,
  tempo_1_seg2									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_179                       									,
  tempo_1_envolvidos_seg2							= x0_2_os_alunos_estao_envolvidos_na_tarefa_180                                                               									,
  tempo_2_seg2									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_181                       									,
  tempo_2_envolvidos_seg2							= x0_2_os_alunos_estao_envolvidos_na_tarefa_182                                                               									,
  tempo_3_seg2									= x0_1_a_professora_promove_uma_atividade_de_aprendizagem_para_a_maioria_dos_alunos_183                       									,
  tempo_3_envolvidos_seg2							= x0_2_os_alunos_estao_envolvidos_na_tarefa_184                                                               									,
  a1_apoio_aprend_1_seg2							= x1_1_a_professora_trata_todos_os_alunos_com_respeito_188                                                                                      ,  
  a1_apoio_aprend_2_seg2							= x1_2_a_professora_usa_uma_linguagem_positiva_com_os_alunos_189                                                                                ,  
  a1_apoio_aprend_3_seg2							= x1_3_a_professora_corresponde_as_necessidades_dos_alunos_190                                                                                  ,  
  a1_apoio_aprend_4_seg2							= x1_4_a_professora_nao_apresenta_preconceitos_de_genero_e_rejeita_estereotipos_de_genero_na_sala_de_aula_191                                   ,  
  a1_expect_posit_1_seg2							= x2_1_a_professora_define_claramente_as_expetativas_de_comportamento_para_as_atividades_na_sala_de_aula_202                                    ,  
  a1_expect_posit_2_seg2							= x2_2_a_professora_reconhece_o_comportamento_positivo_dos_alunos_203                                                                           ,  
  a1_expect_posit_3_seg2							= x2_3_a_professora_redireciona_o_mau_comportamento_e_concentra_se_no_comportamento_esperado_em_vez_de_no_comportamento_indesejado_204          ,  
  a2_licao_clara_1_seg2							= x3_1_a_professora_articula_explicitamente_os_objetivos_da_aula_e_relaciona_as_atividades_da_turma_com_os_objetivos_215                        ,  
  a2_licao_clara_2_seg2							= x3_2_a_explicacao_da_professora_sobre_o_conteudo_e_clara_216                                                                                  ,  
  a2_licao_clara_3_seg2							= x3_3_durante_a_aula_a_professora_estabelece_conexoes_e_relaciona_as_com_outros_conteudos_conhecidos_ou_com_a_vida_quotidiana_dos_alunos_217   ,  
  a2_licao_clara_4_seg2							= x3_4_a_professora_exemplifica_demonstrando_ou_pensando_em_voz_alta_218                                                                        ,  
  a2_compreens_1_seg2								= x4_1_a_professora_faz_perguntas_da_pistas_ou_usa_outras_estrategias_para_determinar_qual_e_o_nivel_de_compreensao_dos_alunos_229              ,  
  a2_compreens_2_seg2								= x4_2_a_professora_monitoriza_a_maioria_dos_alunos_durante_o_trabalho_individual_ou_de_grupo_230                                               ,  
  a2_compreens_3_seg2								= x4_3_a_professora_ajusta_o_ensino_ao_nivel_dos_alunos_231                                                                                     ,  
  a2_coment_const_1_seg2							= x5_1_a_professora_faz_comentarios_especificos_ou_da_pistas_que_ajudam_a_esclarecer_os_equivocos_dos_alunos_242                                ,  
  a2_coment_const_2_seg2							= x5_2_a_professora_comentarios_especificos_ou_da_pistas_que_ajudam_os_alunos_a_identificarem_os_seus_sucessos_243                              ,  
  a2_racioc_crit_1_seg2							= x6_1_a_professora_faz_perguntas_de_resposta_aberta_254                                                                                        ,  
  a2_racioc_crit_2_seg2							= x6_2_a_professora_propoe_atividades_de_raciocinio_255                                                                                         ,  
  a2_racioc_crit_3_seg2							= x6_3_os_alunos_fazem_perguntas_de_resposta_aberta_ou_desenvolvem_atividades_de_raciocinio_256                                                 ,  
  a3_autonomia_1_seg2								= x7_1_a_professora_proporciona_escolhas_aos_alunos_267                                                                                         ,  
  a3_autonomia_2_seg2								= x7_2_a_professora_oferece_oportunidades_aos_alunos_para_assumirem_papeis_na_sala_de_aula_268                                                  ,  
  a3_autonomia_3_seg2								= x7_3_os_alunos_voluntariam_se_para_participar_na_aula_269                                                                                     ,  
  a3_persev_1_seg2								= x8_1_a_professora_reconhece_o_esforco_dos_alunos_280                                                                                          ,  
  a3_persev_2_seg2								= x8_2_a_professora_tem_uma_atitude_positiva_em_relacao_aos_desafios_dos_alunos_281                                                             ,  
  a3_persev_3_seg2								= x8_3_a_professora_incentiva_a_definicao_de_objetivos_282                                                                                      ,  
  a3_cap_soc_colab_1_seg2							= x9_1_a_professora_promove_a_colaboracao_entre_alunos_atraves_da_interacao_entre_colegas_293                                                   ,  
  a3_cap_soc_colab_2_seg2							= x9_2_a_professora_promove_as_capacidades_interpessoais_dos_alunos_294                                                                         ,  
  a3_cap_soc_colab_3_seg2							= x9_3_os_alunos_colaboram_uns_com_os_outros_atraves_da_interacao_entre_eles_295                                                                ,
  meninos_sala									= numero_de_meninos_na_sala_de_aula																												,
  meninas_sala									= numero_de_meninas_na_sala_de_aula																												,
  disciplina										= disciplina																																	,	
  disciplina_matematica							= disciplina_matematica                                                                               											,
  disciplina_portugues							= disciplina_portugues                                                                                											,
  disciplina_ciencias_naturais					= disciplina_ciencias_naturais                                                                        											,
  disciplina_ciencias_sociais						= disciplina_ciencias_sociais                                                                         											,
  disciplina_outra								= disciplina_outra                                                                                    											,
  disciplina_outra_esp							= especifique_a_s_disciplina_s																													,
  possui_lousa									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_quadro_negro_ou_branco																	,
  possui_gis										= qual_desses_recursos_estao_disponivel_na_sala_de_aula_giz_ou_marcadores																		,	
  possui_livros									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_livros																					,
  possui_notebook									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_computador_notebook_tablets																,
  possui_projetor									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_lousa_digital_projector_digital															,
  possui_TV										= qual_desses_recursos_estao_disponivel_na_sala_de_aula_televisao																				,	
  possui_audio									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_radio_ou_equipamento_audio																,
  possui_mat_cienc								= qual_desses_recursos_estao_disponivel_na_sala_de_aula_materiais_de_ciencias																	,	
  possui_mat_pedag								= qual_desses_recursos_estao_disponivel_na_sala_de_aula_mapas_materiais_pedagogicos																,
  possui_jogos									= qual_desses_recursos_estao_disponivel_na_sala_de_aula_brinquedos_jogos_educativos																,
  possui_livros_leitura							= qual_desses_recursos_estao_disponivel_na_sala_de_aula_livros_de_leitura																		,	
  possui_trab_paredes								= qual_desses_recursos_estao_disponivel_na_sala_de_aula_trabalho_e_desenhos_dos_estudantes_nas_paredes											,
  possui_cadeira_todos							= qual_desses_recursos_estao_disponivel_na_sala_de_aula_cadeiras_para_todos_os_alunos															,	
  possui_mesa_todos								= qual_desses_recursos_estao_disponivel_na_sala_de_aula_mesas_ou_mesas_para_todos_os_alunos														) %>%
  
  select(
    escola							,	
    data							,
    classe							,
    professor						,
    observador						,
    observador_outro				,
    professor_presente				,
    professor_sexo					,
    professor_suplente				,
    professor_suplente_sexo			,
    professor_idade					,
    professor_nivel					,
    professor_experiencia			,
    professor_agregado_tamanho		,
    opiniao_prof1 					,
    opiniao_prof2 					,
    opiniao_prof3 					,
    opiniao_prof4 					,
    opiniao_prof5 					,
    opiniao_prof6 					,
    opiniao_prof7 					,
    opiniao_prof8 					,
    opiniao_prof9 					,
    opiniao_prof10					,
    opiniao_prof11					,
    expect_prof_1					,
    expect_prof_2					,
    expect_prof_3					,
    expect_prof_4					,
    expect_prof_5					,
    duracao_seg1					,
    tempo_1_seg1					,
    tempo_1_envolvidos_seg1			,
    tempo_2_seg1					,
    tempo_2_envolvidos_seg1			,
    tempo_3_seg1					,
    tempo_3_envolvidos_seg1			,
    a1_apoio_aprend_1_seg1			,
    a1_apoio_aprend_2_seg1			,
    a1_apoio_aprend_3_seg1			,
    a1_apoio_aprend_4_seg1			,
    a1_expect_posit_1_seg1			,
    a1_expect_posit_2_seg1			,
    a1_expect_posit_3_seg1			,
    a2_licao_clara_1_seg1			,
    a2_licao_clara_2_seg1			,
    a2_licao_clara_3_seg1			,
    a2_licao_clara_4_seg1			,
    a2_compreens_1_seg1				,
    a2_compreens_2_seg1				,
    a2_compreens_3_seg1				,
    a2_coment_const_1_seg1			,
    a2_coment_const_2_seg1			,
    a2_racioc_crit_1_seg1			,
    a2_racioc_crit_2_seg1			,
    a2_racioc_crit_3_seg1			,
    a3_autonomia_1_seg1				,
    a3_autonomia_2_seg1				,
    a3_autonomia_3_seg1				,
    a3_persev_1_seg1				,
    a3_persev_2_seg1				,
    a3_persev_3_seg1				,
    a3_cap_soc_colab_1_seg1			,
    a3_cap_soc_colab_2_seg1			,
    a3_cap_soc_colab_3_seg1			,
    tempo_1_seg2					,
    tempo_1_envolvidos_seg2			,
    tempo_2_seg2					,
    tempo_2_envolvidos_seg2			,
    tempo_3_seg2					,
    tempo_3_envolvidos_seg2			,
    a1_apoio_aprend_1_seg2			,
    a1_apoio_aprend_2_seg2			,
    a1_apoio_aprend_3_seg2			,
    a1_apoio_aprend_4_seg2			,
    a1_expect_posit_1_seg2			,
    a1_expect_posit_2_seg2			,
    a1_expect_posit_3_seg2			,
    a2_licao_clara_1_seg2			,
    a2_licao_clara_2_seg2			,
    a2_licao_clara_3_seg2			,
    a2_licao_clara_4_seg2			,
    a2_compreens_1_seg2				,
    a2_compreens_2_seg2				,
    a2_compreens_3_seg2				,
    a2_coment_const_1_seg2			,
    a2_coment_const_2_seg2			,
    a2_racioc_crit_1_seg2			,
    a2_racioc_crit_2_seg2			,
    a2_racioc_crit_3_seg2			,
    a3_autonomia_1_seg2				,
    a3_autonomia_2_seg2				,
    a3_autonomia_3_seg2				,
    a3_persev_1_seg2				,
    a3_persev_2_seg2				,
    a3_persev_3_seg2				,
    a3_cap_soc_colab_1_seg2			,
    a3_cap_soc_colab_2_seg2			,
    a3_cap_soc_colab_3_seg2			,
    meninos_sala					,
    meninas_sala					,
    disciplina						,
    disciplina_matematica			,
    disciplina_portugues			,
    disciplina_ciencias_naturais	,
    disciplina_ciencias_sociais		,
    disciplina_outra				,
    disciplina_outra_esp			,
    possui_lousa					,
    possui_gis						,
    possui_livros					,
    possui_notebook					,
    possui_projetor					,
    possui_TV						,
    possui_audio					,
    possui_mat_cienc				,
    possui_mat_pedag				,
    possui_jogos					,
    possui_livros_leitura			,
    possui_trab_paredes				,
    possui_cadeira_todos			,
    possui_mesa_todos				) %>% 
  
  relocate(
    escola							,	
    data							,
    classe							,
    professor						,
    observador						,
    observador_outro				,
    professor_presente				,
    professor_sexo					,
    professor_suplente				,
    professor_suplente_sexo			,
    professor_idade					,
    professor_nivel					,
    professor_experiencia			,
    professor_agregado_tamanho		,
    opiniao_prof1 					,
    opiniao_prof2 					,
    opiniao_prof3 					,
    opiniao_prof4 					,
    opiniao_prof5 					,
    opiniao_prof6 					,
    opiniao_prof7 					,
    opiniao_prof8 					,
    opiniao_prof9 					,
    opiniao_prof10					,
    opiniao_prof11					,
    expect_prof_1					,
    expect_prof_2					,
    expect_prof_3					,
    expect_prof_4					,
    expect_prof_5					,
    duracao_seg1					,
    tempo_1_seg1					,
    tempo_1_envolvidos_seg1			,
    tempo_2_seg1					,
    tempo_2_envolvidos_seg1			,
    tempo_3_seg1					,
    tempo_3_envolvidos_seg1			,
    a1_apoio_aprend_1_seg1			,
    a1_apoio_aprend_2_seg1			,
    a1_apoio_aprend_3_seg1			,
    a1_apoio_aprend_4_seg1			,
    a1_expect_posit_1_seg1			,
    a1_expect_posit_2_seg1			,
    a1_expect_posit_3_seg1			,
    a2_licao_clara_1_seg1			,
    a2_licao_clara_2_seg1			,
    a2_licao_clara_3_seg1			,
    a2_licao_clara_4_seg1			,
    a2_compreens_1_seg1				,
    a2_compreens_2_seg1				,
    a2_compreens_3_seg1				,
    a2_coment_const_1_seg1			,
    a2_coment_const_2_seg1			,
    a2_racioc_crit_1_seg1			,
    a2_racioc_crit_2_seg1			,
    a2_racioc_crit_3_seg1			,
    a3_autonomia_1_seg1				,
    a3_autonomia_2_seg1				,
    a3_autonomia_3_seg1				,
    a3_persev_1_seg1				,
    a3_persev_2_seg1				,
    a3_persev_3_seg1				,
    a3_cap_soc_colab_1_seg1			,
    a3_cap_soc_colab_2_seg1			,
    a3_cap_soc_colab_3_seg1			,
    tempo_1_seg2					,
    tempo_1_envolvidos_seg2			,
    tempo_2_seg2					,
    tempo_2_envolvidos_seg2			,
    tempo_3_seg2					,
    tempo_3_envolvidos_seg2			,
    a1_apoio_aprend_1_seg2			,
    a1_apoio_aprend_2_seg2			,
    a1_apoio_aprend_3_seg2			,
    a1_apoio_aprend_4_seg2			,
    a1_expect_posit_1_seg2			,
    a1_expect_posit_2_seg2			,
    a1_expect_posit_3_seg2			,
    a2_licao_clara_1_seg2			,
    a2_licao_clara_2_seg2			,
    a2_licao_clara_3_seg2			,
    a2_licao_clara_4_seg2			,
    a2_compreens_1_seg2				,
    a2_compreens_2_seg2				,
    a2_compreens_3_seg2				,
    a2_coment_const_1_seg2			,
    a2_coment_const_2_seg2			,
    a2_racioc_crit_1_seg2			,
    a2_racioc_crit_2_seg2			,
    a2_racioc_crit_3_seg2			,
    a3_autonomia_1_seg2				,
    a3_autonomia_2_seg2				,
    a3_autonomia_3_seg2				,
    a3_persev_1_seg2				,
    a3_persev_2_seg2				,
    a3_persev_3_seg2				,
    a3_cap_soc_colab_1_seg2			,
    a3_cap_soc_colab_2_seg2			,
    a3_cap_soc_colab_3_seg2			,
    meninos_sala					,
    meninas_sala					,
    disciplina						,
    disciplina_matematica			,
    disciplina_portugues			,
    disciplina_ciencias_naturais	,
    disciplina_ciencias_sociais		,
    disciplina_outra				,
    disciplina_outra_esp			,
    possui_lousa					,
    possui_gis						,
    possui_livros					,
    possui_notebook					,
    possui_projetor					,
    possui_TV						,
    possui_audio					,
    possui_mat_cienc				,
    possui_mat_pedag				,
    possui_jogos					,
    possui_livros_leitura			,
    possui_trab_paredes				,
    possui_cadeira_todos			,
    possui_mesa_todos				)

vetor_vars<- c(
  "duracao"				,
  "tempo_1"	            ,
  "tempo_1_envolvidos"	,
  "tempo_2"	            ,
  "tempo_2_envolvidos"	,
  "tempo_3"	            ,
  "tempo_3_envolvidos"	,
  "a1_apoio_aprend_1"		,
  "a1_apoio_aprend_2"		,
  "a1_apoio_aprend_3"		,
  "a1_apoio_aprend_4"		,
  "a1_expect_posit_1"		,
  "a1_expect_posit_2"		,
  "a1_expect_posit_3"		,
  "a2_licao_clara_1"		,
  "a2_licao_clara_2"		,
  "a2_licao_clara_3"		,
  "a2_licao_clara_4"		,
  "a2_compreens_1"		,
  "a2_compreens_2"		,
  "a2_compreens_3"		,
  "a2_coment_const_1"		,
  "a2_coment_const_2"		,
  "a2_racioc_crit_1"		,
  "a2_racioc_crit_2"		,
  "a2_racioc_crit_3"		,
  "a3_autonomia_1"		,
  "a3_autonomia_2"		,
  "a3_autonomia_3"		,
  "a3_persev_1"	        ,
  "a3_persev_2"	        ,
  "a3_persev_3"	        ,
  "a3_cap_soc_colab_1"	,
  "a3_cap_soc_colab_2"	,
  "a3_cap_soc_colab_3"	)


  
# Pivotando

base_teach3 <- base_teach2 %>%   
  pivot_longer(
    cols = c(duracao_seg1:a3_cap_soc_colab_3_seg2),
    names_to = c('.value', 'segmento'), 
    names_pattern = '(.*)_seg(\\d+)'
    )


saveRDS(base_teach3, file=paste0(tmp, "teach_limpo.RData"))
